+++ include/inc_tmpl/content/cnt32.inc.php
@@ -259,7 +259,7 @@ if(is_array($tmpllist) && count($tmpllist)) {
                   $tab_fieldgroup['fields'][$custom_field]['step'] = ceil($tab_fieldgroup['fields'][$custom_field]['step']);
                 } else {
                   $tab_fieldgroup['fields'][$custom_field]['step'] = floatval($tab_fieldgroup['fields'][$custom_field]['step']);
-                  $tab_fieldgroup['fields'][$custom_field]['step'] = rtrim(number_format($tab_fieldgroup['fields'][$custom_field]['step'], 14 - log10($tab_fieldgroup['fields'][$custom_field]['step'])), '0');
+                  $tab_fieldgroup['fields'][$custom_field]['step'] = rtrim(number_format($tab_fieldgroup['fields'][$custom_field]['step'], (int) (14 - log10($tab_fieldgroup['fields'][$custom_field]['step']))), '0');
                 }
                 echo $tab_fieldgroup['fields'][$custom_field]['step'];
               ?>"
@@ -474,7 +474,7 @@ function addNewTab(pos) {
                                         $tab_fieldgroup['fields'][$custom_field]['step'] = ceil($tab_fieldgroup['fields'][$custom_field]['step']);
                                     } else {
                                         $tab_fieldgroup['fields'][$custom_field]['step'] = floatval($tab_fieldgroup['fields'][$custom_field]['step']);
-                                        $tab_fieldgroup['fields'][$custom_field]['step'] = rtrim(number_format($tab_fieldgroup['fields'][$custom_field]['step'], 14 - log10($tab_fieldgroup['fields'][$custom_field]['step'])), '0');
+                                        $tab_fieldgroup['fields'][$custom_field]['step'] = rtrim(number_format($tab_fieldgroup['fields'][$custom_field]['step'], (int) (14 - log10($tab_fieldgroup['fields'][$custom_field]['step']))), '0');
                                     }
                                     echo $tab_fieldgroup['fields'][$custom_field]['step'];
                                     ?>"
